requirements.txt
requests==2.31.0
python-telegram-bot==20.3
python-dotenv==1.0.0
pymongo[srv]==4.5.0
certifi